Portfolios

Cholesky-variate distributions are distributions whose variate forms are CholeskyVariate. This means each draw is a factorization of a positive-definite matrix of type LinearAlgebra.Cholesky (the object produced by the function LinearAlgebra.cholesky applied to a dense positive-definite matrix.)

Equal Weights

MPT.equal_weights — Function

equal_weights

\[f(x) = \frac{1}{\pi \sqrt{(x - a) (b - x)}}, \quad x \in [a, b]\]

Long Only Minimum Variance

MPT.long_only_minimum_variance — Function

long_only_minimum_variance

\[f(x) = \frac{1}{\pi \sqrt{(x - a) (b - x)}}, \quad x \in [a, b]\]

Most Diversified Portfolio

MPT.mdp — Function

mdp

\[egin{aligned} \textrm{maximize}_{\vec{w}} \quad \frac{\vec{w}^T \vec{\sigma}}{\sqrt{\vec{w}^T \Sigma \vec{w}}}, \ \textrm{subject to} \quad \vec{1} \vec{w} = 1. nd{aligned}\]

External links

Choueifaty, Y., Froidure, T., & Reynier, J. (2013). Properties of the most diversified portfolio, Journal of investment strategies, 2(2), 49-70, doi: 10.2139/ssrn.1895459

Random Weights

MPT.random_weights — Function

random_weights

\[f(x) = \frac{1}{\pi \sqrt{(x - a) (b - x)}}, \quad x \in [a, b]\]

Index